requirements.txt
customtkinter==5.2.0
PyMuPDF==1.23.3
python_docx==0.8.11
scikit_learn==1.3.0
torch==2.0.1
transformers==4.33.1
